+++ conformity/opencv/tests.h
@@ -21,5 +21,6 @@ std::vector<test>& float_conversion_tests_get();
 std::vector<test>& resize_tests_get();
 std::vector<test>& scale_tests_get();
 std::vector<test>& min_max_tests_get();
+std::vector<test>& in_range_tests_get();
 
 #endif  // KLEIDICV_OPENCV_CONFORMITY_TESTS_H_