+++ scripts/ci-opencv.sh
@@ -67,6 +67,7 @@ CORE_TEST_PATTERNS=(
   '*Core_minMaxIdx*'
   '*Core_Array*'
   '*Compare*'
+  '*Core_InRange/*'
 )
 CORE_TEST_PATTERNS_STR="$(join_strings_with_colon "${CORE_TEST_PATTERNS[*]}")"
 ../../../conformity/opencv_kleidicv/bin/opencv_test_core \